code clean up + reformat
