+++ packages/ui/core-components/src/lib/unsorted/viz/map/components/Points.svelte
@@ -213,63 +213,63 @@
 
 	let values, minValue, maxValue, colorScale, sizeExtents, maxData, maxSizeSq;
 
-	/** @type {string[]} */
-	let backupColors = [
-		'red',
-		'blue',
-		'green',
-		'purple',
-		'orange',
-		'pink',
-		'brown',
-		'teal',
-		'cyan',
-		'black',
-		'magenta',
-		'navy',
-		'olive',
-		'lavender',
-		'crimson',
-		'turquoise',
-		'beige',
-		'aqua',
-		'coral',
-		'gold',
-		'silver',
-		'indigo',
-		'violet',
-		'khaki',
-		'plum',
-		'salmon',
-		'sienna',
-		'chartreuse',
-		'lavenderblush',
-		'lightblue',
-		'lightcoral',
-		'lightgreen',
-		'lightpink',
-		'lightyellow',
-		'darkred',
-		'darkgreen',
-		'darkblue',
-		'darkviolet',
-		'darkorange',
-		'darkcyan',
-		'fuchsia',
-		'gainsboro',
-		'honeydew',
-		'hotpink',
-		'lightgray',
-		'lightseagreen',
-		'lightsalmon',
-		'lightsteelblue',
-		'mediumvioletred',
-		'mediumseagreen',
-		'peachpuff',
-		'powderblue',
-		'seashell',
-		'thistle'
-	];
+	// /** @type {string[]} */
+	// let backupColors = [
+	// 	'red',
+	// 	'blue',
+	// 	'green',
+	// 	'purple',
+	// 	'orange',
+	// 	'pink',
+	// 	'brown',
+	// 	'teal',
+	// 	'cyan',
+	// 	'black',
+	// 	'magenta',
+	// 	'navy',
+	// 	'olive',
+	// 	'lavender',
+	// 	'crimson',
+	// 	'turquoise',
+	// 	'beige',
+	// 	'aqua',
+	// 	'coral',
+	// 	'gold',
+	// 	'silver',
+	// 	'indigo',
+	// 	'violet',
+	// 	'khaki',
+	// 	'plum',
+	// 	'salmon',
+	// 	'sienna',
+	// 	'chartreuse',
+	// 	'lavenderblush',
+	// 	'lightblue',
+	// 	'lightcoral',
+	// 	'lightgreen',
+	// 	'lightpink',
+	// 	'lightyellow',
+	// 	'darkred',
+	// 	'darkgreen',
+	// 	'darkblue',
+	// 	'darkviolet',
+	// 	'darkorange',
+	// 	'darkcyan',
+	// 	'fuchsia',
+	// 	'gainsboro',
+	// 	'honeydew',
+	// 	'hotpink',
+	// 	'lightgray',
+	// 	'lightseagreen',
+	// 	'lightsalmon',
+	// 	'lightsteelblue',
+	// 	'mediumvioletred',
+	// 	'mediumseagreen',
+	// 	'peachpuff',
+	// 	'powderblue',
+	// 	'seashell',
+	// 	'thistle'
+	// ];
 
 	/**
 	 * Initialize the component.
@@ -280,43 +280,13 @@
 			await data.fetch();
 			checkInputs(data, [lat, long]);
 			values = $data.map((d) => d[value]);
-			colorPalette = colorPalette.map((item) => chroma(item).hex());
-
-			if (legendType === 'category') {
-				let uniqueValues = new Set(values);
-				values = [...uniqueValues];
-				let i = 0;
-
-				while (colorPalette.length < values.length) {
-					if (!backupColors[i]) {
-						throw new Error('No more backup colors available.');
-					}
-
-					// Check if the color is not already in colorCategory
-					if (!colorPalette.includes(chroma(backupColors[i]).hex())) {
-						colorPalette.push(chroma(backupColors[i]).hex());
-						i++;
-					} else {
-						i++;
-					}
-				}
-			} else if (legendType === 'scalar') {
-				values.forEach((value) => {
-					if (typeof value !== 'number' && value !== null) {
-						throw new Error('Scalar legend requires numeric values or null.');
-					}
-					if (typeof value === 'number' && isNaN(value)) {
-						throw new Error('Scalar legend requires valid numeric values.');
-					}
-				});
-			}
 			minValue = Math.min(...values);
 			maxValue = Math.max(...values);
 			//bucket legend
 			//conditonal
 			colorScale = chroma.scale(colorPalette).domain([min ?? minValue, max ?? maxValue]);
-
 			if (legendType) {
+				values = map.handleLegendValues(colorPalette, values, legendType)
 				map.buildLegend(colorPalette, values, minValue, maxValue);
 			}
 
@@ -376,19 +346,7 @@
 		setInputDefault(item, name);
 	}
 
-	function handleColor(item, value) {
-		if (!value) return uiColours.blue700;
-
-		if (!item[value]) return colorPalette[values.indexOf(item[value])] ?? colorScale(item[value]);
-
-		if (item[value]) {
-			if (typeof item[value] === 'string') {
-				return colorPalette[values.indexOf(item[value])];
-			} else {
-				return colorScale(item[value]);
-			}
-		}
-	}
+	const handleFillColor = map.handleFillColor
 </script>
 
 <!-- Additional data.fetch() included in await to trigger reactivity. Should ideally be handled in init() in the future. -->
@@ -399,7 +357,7 @@
 			options={{
 				// kw note:
 				//need to clean this logic
-				fillColor: color ?? handleColor(item, value) ?? colorScale(item[value]).hex(), // Color of the circle
+				fillColor: color ?? handleFillColor(item, value, values, colorPalette, colorScale) ?? colorScale(item[value]).hex(), // Color of the circle
 				radius: sizeCol ? bubbleSize(item[sizeCol]) : size, // Radius of the circle in meters
 				fillOpacity: opacity,
 				opacity: opacity,
